Avances en viaje y monopatin

+++ trabajo-especial/microservice-monopatin/src/main/java/com/grupo08/unicen/microservicemonopatin/entity/Monopatin.java
@@ -22,5 +22,5 @@ public class Monopatin {
     private int x;
     private int y;
     private BigDecimal kmTraveled;
-    private BigDecimal useTime;
+    private Long useTime;
 }
